Correct google maps api key and correct error with wrongly formatting dot-comma in numbers
