+++ Chilicki.Commline/Chilicki.Commline.UserInterface/Chilicki.Commline.UserInterface.csproj
@@ -189,9 +189,10 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="HtmlExtensions\HtmlDayTypeExtensions.cs" />
-    <Compile Include="HtmlExtensions\HtmlLineTypeExtensions.cs" />
-    <Compile Include="HtmlExtensions\HtmlStopTypeExtensions.cs" />
+    <Compile Include="Helpers\ControllerActivators\CultureAwareControllerActivator.cs" />
+    <Compile Include="Helpers\HtmlExtensions\HtmlDayTypeExtensions.cs" />
+    <Compile Include="Helpers\HtmlExtensions\HtmlLineTypeExtensions.cs" />
+    <Compile Include="Helpers\HtmlExtensions\HtmlStopTypeExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\BaseResources.Designer.cs">
       <AutoGen>True</AutoGen>
