+++ Chilicki.Commline/Chilicki.Commline.UserInterface/Views/Shared/_Map.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Chilicki.Commline.Application.Enums
-@using Chilicki.Commline.UserInterface.HtmlExtensions
+@using Chilicki.Commline.UserInterface.Helpers.HtmlExtensions
 @using Chilicki.Commline.UserInterface.Resources
 @using Chilicki.Commline.Infrastructure.Settings
 
@@ -7,7 +7,7 @@
     var settings = TempData["settings"] as CommlineSettings;
 }
 
-<script src="http://maps.google.com/maps/api/js?key=AIzaSyBWGMUQJXHJ6E0oOLMs-4Ps1hNCPo3zcno" type="text/javascript"></script>
+<script src="http://maps.google.com/maps/api/js?key=AIzaSyBbqodoh5PXNm88vymhqxPjRU-gGVxMJEs" type="text/javascript"></script>
 
 <br />
 <div id="map_canvas"></div>
